# docs/openapi.json
{"openapi":"3.0.0","paths":{"/api/auth/staff/login":{"post":{"operationId":"AuthController_staffLogin","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StaffLoginDto"}}}},"responses":{"200":{"description":"Returns access + refresh token pair"},"401":{"description":"Invalid credentials"}},"summary":"Staff login with email and password","tags":["Auth"]}},"/api/auth/patient/login":{"post":{"description":"Client completes Firebase phone OTP flow, sends the resulting ID token. Backend verifies with Firebase Admin SDK, looks up the pre-registered patient by phone number, and returns a Huggi JWT pair.","operationId":"AuthController_patientLogin","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatientLoginDto"}}}},"responses":{"200":{"description":"Returns access + refresh token pair for patient"},"401":{"description":"Invalid Firebase token or phone not registered"},"404":{"description":"No patient registered with this phone number"}},"summary":"Patient login via Firebase OTP phone verification","tags":["Auth"]}},"/api/auth/refresh":{"post":{"operationId":"AuthController_refresh","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenDto"}}}},"responses":{"200":{"description":"New access + refresh token pair"},"401":{"description":"Invalid or expired refresh token"}},"summary":"Rotate refresh token and issue a new token pair","tags":["Auth"]}},"/api/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenDto"}}}},"responses":{"204":{"description":"Logged out successfully"}},"security":[{"bearer":[]}],"summary":"Revoke refresh token on the server and clear session","tags":["Auth"]}},"/api/clinics":{"get":{"operationId":"ClinicController_listAll","parameters":[],"responses":{"200":{"description":"Array of all clinic records"}},"security":[{"bearer":[]}],"summary":"List all clinics (super admin only)","tags":["Clinics"]},"post":{"operationId":"ClinicController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateClinicDto"}}}},"responses":{"201":{"description":"Clinic created"}},"security":[{"bearer":[]}],"summary":"Create a new clinic (super admin only)","tags":["Clinics"]}},"/api/clinics/{clinicId}":{"get":{"operationId":"ClinicController_getById","parameters":[{"name":"clinicId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Clinic with clinicUsers"},"404":{"description":"Clinic not found"}},"security":[{"bearer":[]}],"summary":"Get clinic details including staff roster","tags":["Clinics"]},"patch":{"operationId":"ClinicController_update","parameters":[{"name":"clinicId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateClinicDto"}}}},"responses":{"200":{"description":"Updated clinic"}},"security":[{"bearer":[]}],"summary":"Update clinic details","tags":["Clinics"]},"delete":{"operationId":"ClinicController_delete","parameters":[{"name":"clinicId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Clinic deleted"}},"security":[{"bearer":[]}],"summary":"Delete a clinic (super admin only)","tags":["Clinics"]}},"/api/users":{"get":{"operationId":"UserController_listByClinic","parameters":[{"name":"clinicId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Array of users with their clinic roles"}},"security":[{"bearer":[]}],"summary":"List clinic staff users (admin only)","tags":["Users"]},"post":{"description":"Creates the User row and a ClinicUser row with the specified role.","operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"User created and assigned to clinic"},"409":{"description":"Email or phone already in use"}},"security":[{"bearer":[]}],"summary":"Create a staff user and assign a clinic role","tags":["Users"]}},"/api/users/{userId}":{"get":{"operationId":"UserController_getById","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"User with clinic role info"},"404":{"description":"User not found"}},"security":[{"bearer":[]}],"summary":"Get a single staff user","tags":["Users"]},"patch":{"operationId":"UserController_update","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"Updated user"}},"security":[{"bearer":[]}],"summary":"Update a staff user (name, phone, role)","tags":["Users"]},"delete":{"operationId":"UserController_delete","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"User removed"}},"security":[{"bearer":[]}],"summary":"Remove a staff user from the clinic","tags":["Users"]}},"/api/doctors":{"get":{"operationId":"DoctorController_listByClinic","parameters":[{"name":"clinicId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Array of doctor profiles with user info"}},"security":[{"bearer":[]}],"summary":"List doctors for a clinic","tags":["Doctors"]},"post":{"operationId":"DoctorController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDoctorDto"}}}},"responses":{"201":{"description":"Doctor profile created"}},"security":[{"bearer":[]}],"summary":"Create a doctor profile for an existing user (admin only)","tags":["Doctors"]}},"/api/doctors/{doctorId}":{"get":{"operationId":"DoctorController_getById","parameters":[{"name":"doctorId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Doctor profile with user info"},"404":{"description":"Doctor not found"}},"security":[{"bearer":[]}],"summary":"Get a single doctor profile","tags":["Doctors"]},"patch":{"operationId":"DoctorController_update","parameters":[{"name":"doctorId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateDoctorDto"}}}},"responses":{"200":{"description":"Updated doctor profile"}},"security":[{"bearer":[]}],"summary":"Update doctor specialization, duration, or availability (admin only)","tags":["Doctors"]},"delete":{"operationId":"DoctorController_delete","parameters":[{"name":"doctorId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Doctor deleted"}},"security":[{"bearer":[]}],"summary":"Delete a doctor profile (admin only)","tags":["Doctors"]}},"/api/patients/me":{"get":{"description":"Patient-only endpoint. Returns the authenticated patient record with user info.","operationId":"PatientController_getMe","parameters":[],"responses":{"200":{"description":"Patient's own profile"},"403":{"description":"Not a patient token"}},"security":[{"bearer":[]}],"summary":"Get authenticated patient's own profile","tags":["Patients"]}},"/api/patients":{"get":{"operationId":"PatientController_listByClinic","parameters":[{"name":"clinicId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Array of patient records with user info"}},"security":[{"bearer":[]}],"summary":"List active patients for a clinic","tags":["Patients"]},"post":{"description":"Creates a User row and a Patient row atomically. Phone must be unique.","operationId":"PatientController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePatientDto"}}}},"responses":{"201":{"description":"Patient registered"},"409":{"description":"Phone number already registered"}},"security":[{"bearer":[]}],"summary":"Register a new patient","tags":["Patients"]}},"/api/patients/{patientId}":{"get":{"operationId":"PatientController_getById","parameters":[{"name":"patientId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Patient with user info"},"404":{"description":"Patient not found"}},"security":[{"bearer":[]}],"summary":"Get a single patient record","tags":["Patients"]},"patch":{"operationId":"PatientController_update","parameters":[{"name":"patientId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePatientDto"}}}},"responses":{"200":{"description":"Updated patient"}},"security":[{"bearer":[]}],"summary":"Update patient notes or active status","tags":["Patients"]},"delete":{"operationId":"PatientController_delete","parameters":[{"name":"patientId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Patient deactivated"}},"security":[{"bearer":[]}],"summary":"Soft-delete a patient (sets isActive=false, preserves history)","tags":["Patients"]}},"/api/appointments":{"get":{"operationId":"AppointmentController_listByClinic","parameters":[{"name":"clinicId","required":true,"in":"query","description":"Clinic UUID","schema":{"type":"string"}}],"responses":{"200":{"description":"Array of appointment records"}},"security":[{"bearer":[]}],"summary":"List all appointments for a clinic (staff only)","tags":["Appointments"]},"post":{"operationId":"AppointmentController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAppointmentDto"}}}},"responses":{"201":{"description":"Appointment created and added to queue"},"409":{"description":"Doctor already booked at this time, or time is in the past"}},"security":[{"bearer":[]}],"summary":"Book a new appointment (staff only)","tags":["Appointments"]}},"/api/appointments/mine":{"get":{"description":"Patient-only endpoint. Requires a JWT issued via POST /auth/patient/login.","operationId":"AppointmentController_mine","parameters":[],"responses":{"200":{"description":"Array of the patient's appointments with doctor details"},"403":{"description":"Not a patient token"}},"security":[{"bearer":[]}],"summary":"Get authenticated patient's own appointments","tags":["Appointments"]}},"/api/appointments/by-patient/{patientId}":{"get":{"operationId":"AppointmentController_getByPatient","parameters":[{"name":"patientId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Array of appointments for the patient"}},"security":[{"bearer":[]}],"summary":"Get appointments for a specific patient (staff only)","tags":["Appointments"]}},"/api/appointments/{appointmentId}":{"get":{"operationId":"AppointmentController_getById","parameters":[{"name":"appointmentId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Appointment record"},"404":{"description":"Appointment not found"}},"security":[{"bearer":[]}],"summary":"Get a single appointment by ID (staff only)","tags":["Appointments"]},"patch":{"operationId":"AppointmentController_update","parameters":[{"name":"appointmentId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAppointmentDto"}}}},"responses":{"200":{"description":"Updated appointment"}},"security":[{"bearer":[]}],"summary":"Update appointment time, status, or notes (staff only)","tags":["Appointments"]},"delete":{"operationId":"AppointmentController_delete","parameters":[{"name":"appointmentId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Appointment deleted"}},"security":[{"bearer":[]}],"summary":"Hard-delete an appointment and its queue entry (admin only)","tags":["Appointments"]}},"/api/appointments/{appointmentId}/cancel":{"post":{"operationId":"AppointmentController_cancel","parameters":[{"name":"appointmentId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Appointment cancelled"}},"security":[{"bearer":[]}],"summary":"Cancel an appointment and remove its queue entry (staff only)","tags":["Appointments"]}},"/api/queue":{"get":{"operationId":"QueueController_listByClinic","parameters":[{"name":"clinicId","required":true,"in":"query","schema":{"type":"string"}},{"name":"status","required":false,"in":"query","schema":{"enum":["waiting","called","in_consultation","done","no_show"],"type":"string"}}],"responses":{"200":{"description":"Queue entries with nested appointment, patient, and doctor data"}},"security":[{"bearer":[]}],"summary":"List clinic queue entries, optionally filtered by status (staff only)","tags":["Queue"]}},"/api/queue/my-position":{"get":{"description":"Patient-only endpoint. Returns the active queue entry (waiting/called/in_consultation) for the patient's most recent appointment today.","operationId":"QueueController_myPosition","parameters":[],"responses":{"200":{"description":"Queue position and status, or null if not in queue"},"403":{"description":"Not a patient token"}},"security":[{"bearer":[]}],"summary":"Get authenticated patient's current queue position","tags":["Queue"]}},"/api/queue/{queueId}":{"get":{"operationId":"QueueController_getById","parameters":[{"name":"queueId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Queue entry"},"404":{"description":"Queue entry not found"}},"security":[{"bearer":[]}],"summary":"Get a single queue entry by ID (staff only)","tags":["Queue"]},"patch":{"operationId":"QueueController_update","parameters":[{"name":"queueId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateQueueDto"}}}},"responses":{"200":{"description":"Updated queue entry"},"400":{"description":"Invalid status value"}},"security":[{"bearer":[]}],"summary":"Update queue entry status (staff only)","tags":["Queue"]},"delete":{"operationId":"QueueController_delete","parameters":[{"name":"queueId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Queue entry deleted"}},"security":[{"bearer":[]}],"summary":"Delete a queue entry and recompact positions (admin only)","tags":["Queue"]}}},"info":{"title":"Huggi Hospital Queue API","description":"REST API for Huggi Super App — clinic queue and appointment management for Indian clinics.","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"StaffLoginDto":{"type":"object","properties":{"email":{"type":"string","example":"admin@clinic.com","description":"Staff email address"},"password":{"type":"string","example":"secret123","minLength":8,"description":"Staff password"}},"required":["email","password"]},"PatientLoginDto":{"type":"object","properties":{"idToken":{"type":"string","description":"Firebase ID token obtained after successful OTP verification on the client","example":"eyJhbGciOiJSUzI1NiIsImtpZCI6..."}},"required":["idToken"]},"RefreshTokenDto":{"type":"object","properties":{"refreshToken":{"type":"string","description":"Opaque refresh token received from login or previous refresh","minLength":32}},"required":["refreshToken"]},"CreateClinicDto":{"type":"object","properties":{"name":{"type":"string","example":"Sunrise Clinic","minLength":3},"address":{"type":"string","example":"12, MG Road, Pune, Maharashtra 411001","minLength":5},"phone":{"type":"string","example":"9876543210","minLength":10,"description":"10-digit clinic phone number"}},"required":["name","address","phone"]},"UpdateClinicDto":{"type":"object","properties":{"name":{"type":"string","example":"Sunrise Super Clinic"},"address":{"type":"string","example":"45, FC Road, Pune, Maharashtra 411004"},"phone":{"type":"string","example":"9876543211"}}},"CreateUserDto":{"type":"object","properties":{"email":{"type":"string","example":"nurse.priya@clinic.com"},"password":{"type":"string","example":"secret123","minLength":8},"name":{"type":"string","example":"Priya Sharma","minLength":2},"phone":{"type":"string","example":"9876500001"},"clinicId":{"type":"string","description":"UUID of the clinic to assign this staff member to"},"role":{"type":"string","example":"NURSE","description":"Staff role: CLINIC_ADMIN | RECEPTIONIST | NURSE | DOCTOR | CARE_TAKER | SECURITY | BILLING_STAFF"}},"required":["email","password","name","clinicId","role"]},"UpdateUserDto":{"type":"object","properties":{"email":{"type":"string","example":"priya.new@clinic.com"},"name":{"type":"string","example":"Priya Sharma Gupta","minLength":2},"phone":{"type":"string","example":"9876500002"}}},"CreateDoctorDto":{"type":"object","properties":{"userId":{"type":"string","description":"UUID of the existing User account for this doctor"},"clinicId":{"type":"string","description":"UUID of the clinic this doctor belongs to"},"specialization":{"type":"string","example":"Cardiology","default":"General"},"consultationDuration":{"type":"number","example":30,"minimum":15,"description":"Slot duration in minutes"}},"required":["userId","clinicId"]},"UpdateDoctorDto":{"type":"object","properties":{"specialization":{"type":"string","example":"Neurology"},"consultationDuration":{"type":"number","example":20,"minimum":15},"isAvailable":{"type":"boolean","example":true,"description":"Whether the doctor is accepting patients today"}}},"CreatePatientDto":{"type":"object","properties":{"name":{"type":"string","example":"Ramesh Kumar","minLength":2},"phone":{"type":"string","example":"9876543210","minLength":10,"description":"10-digit mobile number"},"clinicId":{"type":"string","description":"UUID of the clinic this patient is registered at"},"notes":{"type":"string","example":"Diabetic, allergic to penicillin"}},"required":["name","phone","clinicId"]},"UpdatePatientDto":{"type":"object","properties":{"notes":{"type":"string","example":"Hypertension, regular medication"},"isActive":{"type":"boolean","example":true}}},"CreateAppointmentDto":{"type":"object","properties":{"clinicId":{"type":"string","description":"UUID of the clinic"},"patientId":{"type":"string","description":"UUID of the patient"},"doctorId":{"type":"string","description":"UUID of the doctor"},"appointmentTime":{"type":"string","example":"2026-06-15T10:30:00.000Z","description":"ISO 8601 appointment datetime"},"notes":{"type":"string","example":"Follow-up for hypertension"}},"required":["clinicId","patientId","doctorId","appointmentTime"]},"UpdateAppointmentDto":{"type":"object","properties":{"appointmentTime":{"type":"string","example":"2026-06-15T11:00:00.000Z"},"status":{"type":"string","example":"booked","description":"booked | done | no_show"},"notes":{"type":"string","example":"Patient rescheduled"}}},"UpdateQueueDto":{"type":"object","properties":{"status":{"type":"string","example":"called","description":"New queue status. Allowed: waiting | called | in_consultation | done | no_show","enum":["waiting","called","in_consultation","done","no_show"]},"position":{"type":"number","example":2,"minimum":0}},"required":["status"]}}}}